Rename file from [mianPage.dart] to [mainPage.dart]
